style : main 메서드를 다시 클래스로 만들기
